made new Component AddNewEXercise, code refactor
