+++ src/js/application/exercises.js
@@ -1,25 +1,17 @@
 import React, {useEffect, useState} from 'react';
 import firebase from "firebase";
-import CloseX from "../components/closeX";
+import AddNewExercise from "./AddNewExercise";
 
 function Exercises() {
   const [exercises, setExercises] = useState([])
   const [isLoaded, setLoaded] = useState(false)
   const [isAddNewExVisible, setAddNewExVisible] = useState(false)
   
-  const [addNewExercise, setNewExercise] = useState({
-    id: null,
-    name: '',
-    difficulty: '',
-    type: '',
-    muscleGroup: '',
-    img: '',
-    description: ''
-  })
-  
-  const handleAddOn = () => setAddNewExVisible(true)
-  const handleAddOff = () => setAddNewExVisible(false)
-  
+  const handleAddNewExVisible = (value) => {
+    setAddNewExVisible(value)
+  }
+  const handleOn = () => handleAddNewExVisible(true)
+
   
   const rootRef = firebase.database().ref();
   const exercisesRef = rootRef.child('exercises')
@@ -45,30 +37,7 @@ function Exercises() {
   }
   
   // console.log(typeof addNewExercise);
-  const handleAddNewEx = (event) => {
-    const {name, value} = event.target;
-    setNewExercise(prevState => {
-      return {
-        ...prevState,
-        [name]: value
-      }
-    })
-  }
   
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    const exerciseRef = firebase.database().ref('exercises').push();
-    const newEx = addNewExercise;
-    exerciseRef.set(newEx);
-    handleAddOff();
-  }
-  
-  const isInvalid =
-    addNewExercise.name === '' ||
-    addNewExercise.description === '' ||
-    addNewExercise.difficulty === '' ||
-    addNewExercise.type === '' ||
-    addNewExercise.muscleGroup === '';
   
   return (
     <section className={'exercises'}>
@@ -92,47 +61,14 @@ function Exercises() {
           </form>
         </div>
         <div className={'exercises__display'}>
-          <div onClick={handleAddOn} className={'exercises__box exercises__box-add'}> Add exercise</div>
+          <div onClick={handleOn} className={'exercises__box exercises__box-add'}> Add exercise</div>
           {exercises.map(el => {
             return (
               <div className={'exercises__box'} key={el.id}> {el.name}</div>
             )
           })}
         </div>
-        
-        
-        <div className={isAddNewExVisible ? 'addNewExercise' : 'addNewExercise hidden'}>
-          <div className={'board addNewExercise__board'}>
-            <CloseX onClick={handleAddOff}/>
-            <h2 className={'addNewExercise__title'}>Add new exercise</h2>
-            <form className={'addNewExercise__form'} onSubmit={handleSubmit}>
-              
-              <input type={'text'} name={'name'} value={addNewExercise.name} onChange={handleAddNewEx}/>
-              
-              <input type={'text'} name={'description'} value={addNewExercise.description} onChange={handleAddNewEx}/>
-              
-              <select name={'difficulty'} placeholder={'difficulty'} value={addNewExercise.difficulty}
-                      onChange={handleAddNewEx}>
-                <option value={'easy'}>Easy</option>
-                <option value={'medium'}>Medium</option>
-                <option value={'hard'}>Hard</option>
-              </select>
-              
-              <select name={'muscleGroup'} placeholder={'muscle group'} value={addNewExercise.muscleGroup}
-                      onChange={handleAddNewEx}>
-                <option value={'upper body'}>Upper body</option>
-                <option value={'lower body'}>Lower body</option>
-              </select>
-              
-              <select name={'type'} placeholder={'type'} value={addNewExercise.type} onChange={handleAddNewEx}>
-                <option value={'indoor'}>Indoor</option>
-                <option value={'outdoor'}>Outdoor</option>
-              </select>
-              <button type={'submit'} disabled={isInvalid}>Add exercise</button>
-            </form>
-          </div>
-        
-        </div>
+        <AddNewExercise isAddNewExVisible={isAddNewExVisible} handleAddNewExVisible={handleAddNewExVisible}/>
       </div>
     </section>
   );
